+++ clicks/digipot14/changelog.md
@@ -1,4 +1,4 @@
 ## Changelog
 
-### Version 2.1.0.3
+### Version 2.1.0.2
  - Initial release